+++ examples/deterministic/ripple/RippleEffect.hs
@@ -16,14 +16,14 @@ Module for Ripple Effect Analysis
 module RippleEffect where
 
 -- Original
-import              ExchangeAlgebraJournal
+import              ExchangeAlgebra.Journal
 import qualified    ExchangeAlgebra.Journal  as EJ
 import qualified    ExchangeAlgebra.Journal.Transfer as EJT
 import              ExchangeAlgebra.Journal.Transfer ((.->)
                                                         ,(|%))
 
 import qualified    ExchangeAlgebra.Simulate as ES
-import qualified    ExchangeAlgebra.Simulate
+import              ExchangeAlgebra.Simulate
 import qualified    ExchangeAlgebra.Simulate.Visualize as ESV