+++ examples/deterministic/ripple/rippleWithStockMultiSeeds.hs
@@ -18,14 +18,15 @@ Final demand increases at a specified time, and the impact is measured.
 
 -- Original
 import              RippleEffect
-import              ExchangeAlgebraJournal
+import              ExchangeAlgebra.Journal
 import qualified    ExchangeAlgebra.Journal  as EJ
 import qualified    ExchangeAlgebra.Journal.Transfer as EJT
 import              ExchangeAlgebra.Journal.Transfer ((.->)
                                                         ,(|%))
 
 import qualified    ExchangeAlgebra.Simulate as ES
-import qualified    ExchangeAlgebra.Simulate
+import              ExchangeAlgebra.Simulate
+import              ExchangeAlgebra.Write
 import qualified    ExchangeAlgebra.Simulate.Visualize as ESV