+++ examples/optimization/CGE/stdcge.hs
@@ -15,14 +15,15 @@
 
 -- Original
 import              CGE
-import              ExchangeAlgebraJournal
+import              ExchangeAlgebra.Journal
 import qualified    ExchangeAlgebra.Journal  as EJ
 import qualified    ExchangeAlgebra.Journal.Transfer as EJT
 import              ExchangeAlgebra.Journal.Transfer ((.->)
                                                         ,(|%))
 
 import qualified    ExchangeAlgebra.Simulate as ES
-import qualified    ExchangeAlgebra.Simulate
+import              ExchangeAlgebra.Simulate
+import              ExchangeAlgebra.Write
 import qualified    ExchangeAlgebra.Simulate.Visualize as ESV
 
 -- for python visualization