+++ test/doctests.hs
@@ -4,5 +4,4 @@ import Test.DocTest
 
 main :: IO ()
 main = doctest  [ "-isrc"
-                , "src/ExchangeAlgebra.hs"
-                , "src/ExchangeAlgebraJournal.hs"]
+                , "src/ExchangeAlgebra.hs"]